+++ illuin_inject/bindings/__init__.py
@@ -2,9 +2,10 @@ from .abstract_module import AbstractModule
 from .binding import Binding
 from .binding_registry import BindingRegistry
 from .binding_to_provider_adapter import BindingToProviderAdapter
-from .class_binding import ClassBinding, ClassBindingToProviderAdapter, FromClassProvider
+from .class_binding import ClassBinding, ClassBindingToProviderAdapter
 from .factory_binding import FactoryBinding, FactoryBindingToProviderAdapter, FromFactoryProvider
 from .instance_binding import FromInstanceProvider, InstanceBinding, InstanceBindingToProviderAdapter
 from .module import Module
 from .multi_binding import ItemBinding, ListProvider, MultiBinding, MultiBindingToProviderAdapter
 from .private_module import PrivateModule
+from .self_binding import FromClassProvider, SelfBinding, SelfBindingToProviderAdapter