+++ illuin_inject/bindings/class_binding/__init__.py
@@ -1,3 +1,2 @@
 from .class_binding import ClassBinding
 from .class_binding_to_provider_adapter import ClassBindingToProviderAdapter
-from .from_class_provider import FromClassProvider