+++ illuin_inject/bindings/self_binding/self_binding_to_provider_adapter.py
@@ -1,0 +1,63 @@
+from inspect import Parameter, signature
+from typing import Dict, List, Type
+
+from illuin_inject.bindings.binding import Binding
+from illuin_inject.bindings.binding_to_provider_adapter import BindingToProviderAdapter
+from illuin_inject.bindings.instance_binding import FromInstanceProvider
+from illuin_inject.exceptions import NoBindingFound, NonInjectableTypeError
+from illuin_inject.injection_state import InjectionState
+from illuin_inject.provider import Provider
+from illuin_inject.target import Target
+from illuin_inject.type_checker import TypeChecker
+from illuin_inject.typings import InjectedT
+from .from_class_provider import FromClassProvider
+from .self_binding import SelfBinding
+
+
+# pylint: disable=no-self-use, unused-argument
+class SelfBindingToProviderAdapter(BindingToProviderAdapter[SelfBinding]):
+    """Creates a Provider from a SelfBinding."""
+
+    def accept(self, binding: Binding[InjectedT], state: InjectionState) -> bool:
+        return isinstance(binding, SelfBinding)
+
+    def create(self, binding: SelfBinding[InjectedT], state: InjectionState) -> Provider:
+        parameters = signature(binding.target_type.__init__).parameters
+        args: List[Provider] = []
+        kwargs: Dict[str, Provider] = {}
+        # Ignore 'self'
+        for parameter in list(parameters.values())[1:]:
+            # Ignore '*args' and '**kwargs'
+            if parameter.kind in [Parameter.VAR_POSITIONAL, Parameter.VAR_KEYWORD]:
+                continue
+
+            parameter_provider = self._get_parameter_provider(parameter, binding.target_type, state)
+            if parameter.kind == Parameter.KEYWORD_ONLY:
+                kwargs[parameter.name] = parameter_provider
+            else:
+                args.append(parameter_provider)
+        unscoped_provider = FromClassProvider(binding.target_type, args, kwargs)
+        try:
+            scope_provider = state.provider_creator.get_provider(Target(binding.scope), state)
+        except NoBindingFound:
+            raise NonInjectableTypeError(f"Could not create a provider for {binding!r}: they are no bindings for"
+                                         f" {binding.scope.__name__!r}")
+        return scope_provider.get().get_scoped_provider(unscoped_provider)
+
+    @staticmethod
+    def _get_parameter_provider(parameter: Parameter,
+                                current_class: Type,
+                                state: InjectionState) -> Provider:
+        if parameter.annotation is not Parameter.empty:
+            if TypeChecker.is_annotated(parameter.annotation):
+                target = Target(parameter.annotation.original_type, parameter.annotation.annotation)
+            else:
+                target = Target(parameter.annotation, None)
+            try:
+                return state.provider_creator.get_provider(target, state)
+            except NoBindingFound:
+                pass
+        if parameter.default is not Parameter.empty:
+            return FromInstanceProvider(parameter.default)
+        raise NonInjectableTypeError(f"Could not find a binding or a default value for {parameter.name}: "
+                                     f"{parameter.annotation} required by {current_class}")